requirements.txt
google-generativeai
chromadb
python-dotenv
httpx
mcp
google-adk
